test: add happy path E2E, encoding and decoding unit tests for getFungibleTokenInfo() precompile

+++ hedera-node/src/main/java/com/hedera/services/store/contracts/precompile/codec/FungibleTokenInfo.java
@@ -15,4 +15,4 @@
  */
 package com.hedera.services.store.contracts.precompile.codec;
 
-public record FungibleTokenInfo(TokenInfo tokenInfo, long decimals) {}
+public record FungibleTokenInfo(TokenInfo tokenInfo, int decimals) {}
